Added setup for transactional testing

+++ src/main/java/com/sparta/academy/mfix_mongodb_api/controller/UserController.java
@@ -5,9 +5,12 @@ import com.sparta.academy.mfix_mongodb_api.exceptions.IDNotFoundException;
 import com.sparta.academy.mfix_mongodb_api.logging.LoggerSingleton;
 import com.sparta.academy.mfix_mongodb_api.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.logging.Logger;
 
 @RestController
@@ -34,8 +37,12 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public void deleteUser(@PathVariable String id) {
-        userRepository.deleteById(id);
+    public ResponseEntity<String> deleteUser(@PathVariable String id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).body("User with ID " + id + " has been deleted Successfully!");
+        }
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("No User with ID " + id);
     }
 
     @PostMapping("/users")
